contract-event-listener/requirements.txt
attrs==19.3.0
base58==2.0.0
bitarray==1.2.2
boto3==1.13.18
botocore==1.16.18
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
cytoolz==0.10.1
docutils==0.15.2
eth-abi==2.1.1
eth-account==0.5.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.5
flake8==3.8.2
hexbytes==0.2.0
idna==2.9
ipfshttpclient==0.4.13.2
jmespath==0.10.0
jsonschema==3.2.0
lru-dict==1.1.6
marshmallow==3.6.0
mccabe==0.6.1
more-itertools==8.3.0
multiaddr==0.0.9
netaddr==0.7.19
packaging==20.4
parsimonious==0.8.1
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pycodestyle==2.6.0
pycryptodome==3.9.7
pyflakes==2.2.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
rlp==1.2.0
s3transfer==0.3.3
sentry-sdk==0.14.4
six==1.15.0
toolz==0.10.0
urllib3==1.25.9
varint==1.0.2
wcwidth==0.1.9
web3==5.12.1
websockets==8.1
